+++ Python-Server/app/image_tagging/tensorflow_tagging/tensorflow_tagging.py
@@ -3,11 +3,12 @@ from __future__ import division
 from __future__ import print_function
 
 import numpy as np
+from PIL import UnidentifiedImageError
 
 from tensorflow.python.keras.applications.inception_v3 import InceptionV3, preprocess_input, decode_predictions
 from tensorflow.python.keras.preprocessing import image
 import os
-from utils.utils import download_image_from_url, createFolder
+from server_utils.utils import download_image_from_url, createFolder
 
 
 class TensorFlowTag:
@@ -22,24 +23,26 @@ class TensorFlowTag:
         else:
             file = download_image_from_url(image_url)
 
-        img = image.load_img(file, target_size=(299, 299))
-        x = image.img_to_array(img)
-        x = np.expand_dims(x, axis=0)
-        x = preprocess_input(x)
+        try:
+            img = image.load_img(file, target_size=(299, 299))
+            x = image.img_to_array(img)
+            x = np.expand_dims(x, axis=0)
+            x = preprocess_input(x)
 
-        preds = self.model.predict(x)
-        prediction_decoded = decode_predictions(preds, top=20)[0]
+            preds = self.model.predict(x)
+            prediction_decoded = decode_predictions(preds, top=20)[0]
 
-        print('Predicted:', prediction_decoded)
+            print('Predicted:', prediction_decoded)
 
-        tags = []
-        for element in prediction_decoded:
-            if element[2] > treshold:
-                tags.append(element[1])
+            tags = []
+            for element in prediction_decoded:
+                if element[2] > treshold:
+                    tags.append(element[1])
 
-        if self.delete_after_tagging:
-            os.remove(file)
+            if self.delete_after_tagging:
+                os.remove(file)
 
-        # str1 = ','.join(str(e) for e in tags)
+        except UnidentifiedImageError as e:
+            self.logging.info(str(e))
 
         return tags