+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+## [2.0.5]
+
+### Added 
+
+- Now you can add Video files from web or phone!
+- Started the base for adding a background worker that maintains the database, cleaning duplicated or blank content added. (Coming soon)
+- Tags added from the AI tagging system have a number filter, so they are not all sent at the same time only the 10 first that are more used!
+
+### Fixed
+- More error handling and structure fix to make everything faster.
+- Bug fix https://github.com/elblogbruno/NotionAI-MyMind/issues/20
 
 ## [2.0.4]
 
@@ -122,6 +133,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fixed server issues adding callback capabilities. Before some websites images were not added due to timeouts.
 
+[2.0.5]: https://github.com/elblogbruno/NotionAI-MyMind/releases/tag/2.0.5
 [2.0.4]: https://github.com/elblogbruno/NotionAI-MyMind/releases/tag/2.0.4
 [2.0.3]: https://github.com/elblogbruno/NotionAI-MyMind/releases/tag/2.0.3
 [2.0.2]: https://github.com/elblogbruno/NotionAI-MyMind/releases/tag/2.0.2